Introduce epochs
This commit renames WorkTree to epoch and wraps it in a new
implementation of WorkTree, which can switch between epochs. The new
WorkTree is designed to perform I/O interactions with a GitProvider, and
therefore returns streams to represent async computations in response to
that I/O.
