+++ memo_core/Cargo.toml
@@ -5,6 +5,7 @@ authors = ["Antonio Scandurra <as-cii@github.com>", "Nathan Sobo <nathan@github.
 
 [dependencies]
 lazy_static = "1.0"
+futures = "0.1"
 serde = "1.0"
 serde_derive = "1.0"
 smallvec = "0.6.1"
